refactor builder modules (#249)
This splits the builders to their own files
